询单、订单、客人编辑页面，添加注释，不再在页面直接添加注释后立刻保存,将统计界面的合计和统计数据合并为同一个datatable显示,用户列表页，应该显示用户名而不是登录名

+++ src/com/tourmade/crm/action/AgencyController.java
@@ -130,16 +130,21 @@ public class AgencyController extends BaseSimpleFormController {
 		try {
 			if (null != id && !"".equals(id)) {
 				int agencyId = Integer.parseInt(id);
-				service.deleteAgencyById(agencyId);
-				json.setSuccess(true);
-			} else {
+				String agencyName = service.checkAgencyHaveOrder(agencyId);
+				if("".equals(agencyName) && null!=agencyName){
+					service.deleteAgencyById(agencyId);
+					json.setSuccess(true);
+				}else{
+					//json.setText("有地接社不能删除");
+					json.setSuccess(false);
+				}
+			} else {}
 				json.setSuccess(false);
-			}
+			
 		} catch (Exception e) {
 			json.setSuccess(false);
 			logger.error("AgencyController.doDel() --> " + id + "\n" + e.getMessage());
 		}		
 		return json;
 	}
-
 }
